LPS-184766 Activator & Public Package: move session listener registration into the HttpServletService tracker so that we don't need to expose HttpSessionTrackerUtil
