+++ modules/apps/static/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/osgi/web/wab/extender/internal/adapter/HttpAdapter.java
@@ -5,10 +5,8 @@
 
 package com.liferay.portal.osgi.web.wab.extender.internal.adapter;
 
-import com.liferay.portal.kernel.servlet.PortletSessionListenerManager;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.osgi.web.http.servlet.HttpServletService;
-import com.liferay.portal.osgi.web.http.servlet.HttpSessionTrackerUtil;
 import com.liferay.portal.osgi.web.wab.extender.internal.registration.ServletRegistrationImpl;
 
 import java.lang.reflect.InvocationHandler;
@@ -26,9 +24,6 @@ import javax.servlet.ServletContext;
 import javax.servlet.ServletRegistration;
 import javax.servlet.descriptor.JspConfigDescriptor;
 import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpSession;
-import javax.servlet.http.HttpSessionEvent;
-import javax.servlet.http.HttpSessionListener;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
@@ -94,16 +89,10 @@ public class HttpAdapter {
 
 		_serviceRegistration = bundleContext.registerService(
 			HttpServletService.class, () -> servletConfig, properties);
-
-		PortletSessionListenerManager.addHttpSessionListener(
-			_INVALIDATEHTTPSESSION_LISTENER);
 	}
 
 	@Deactivate
 	protected void deactivate() {
-		PortletSessionListenerManager.removeHttpSessionListener(
-			_INVALIDATEHTTPSESSION_LISTENER);
-
 		_serviceRegistration.unregister();
 
 		_serviceRegistration = null;
@@ -112,22 +101,6 @@ public class HttpAdapter {
 	private static final Class<?>[] _INTERFACES = new Class<?>[] {
 		ServletContext.class
 	};
-
-	private static final HttpSessionListener _INVALIDATEHTTPSESSION_LISTENER =
-		new HttpSessionListener() {
-
-			@Override
-			public void sessionCreated(HttpSessionEvent httpSessionEvent) {
-			}
-
-			@Override
-			public void sessionDestroyed(HttpSessionEvent httpSessionEvent) {
-				HttpSession httpSession = httpSessionEvent.getSession();
-
-				HttpSessionTrackerUtil.invalidate(httpSession.getId());
-			}
-
-		};
 
 	private ServiceRegistration<?> _serviceRegistration;
